backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.3
openai>=1.12.0
python-dotenv==1.0.0
requests==2.31.0
motor>=3.7.1
pymongo>=4.16.0
pyppeteer==1.0.2